move lambda processing test to query builder preset
